Caching and multithreading

+++ src/main/java/com/cla/employeeportaldemo/service/DepartmentServiceImpl.java
@@ -2,6 +2,8 @@ package com.cla.employeeportaldemo.service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
 import javax.transaction.Transactional;
@@ -12,6 +14,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
 import com.cla.employeeportaldemo.controller.EmployeeController;
@@ -21,6 +24,7 @@ import com.cla.employeeportaldemo.dto.EmployeeDepartmentDTO;
 import com.cla.employeeportaldemo.entity.Department;
 import com.cla.employeeportaldemo.entity.Employee;
 import com.cla.employeeportaldemo.entity.EmployeeDepartment;
+import com.cla.employeeportaldemo.exception.IDNotFoundException;
 import com.cla.employeeportaldemo.repository.DepartmentRepository;
 import com.cla.employeeportaldemo.repository.EmployeeDepartmentRepository;
 import com.cla.employeeportaldemo.repository.EmployeeRepository;
@@ -38,29 +42,28 @@ public class DepartmentServiceImpl implements DepartmentService
 	 private EmployeeDepartmentRepository employeeDepartmentRepository;
 	 
 	 private static final Logger logger = LogManager.getLogger(DepartmentServiceImpl.class);
+	
 	 
 	@Transactional
 	@Override
-	@Cacheable(cacheNames = "employeeCache")
 	public List<DepartmentDTO> findAll() 
 	{
-
 		List<DepartmentDTO> departmentDTO = new ArrayList<>();
 		List<Department> department = departmentRepository.findAll();
-		logger.info("Fetching Department details aalong with employee details");
+		logger.info("Fetching Department details aalong with Department details");
+		logger.info("Fetching Department details:  "+Thread.currentThread().getName());
 		department.stream().forEach(dept -> {
 			DepartmentDTO deptdto = mapEntityToDto(dept);
 			departmentDTO.add(deptdto);
 		});
-
 		return departmentDTO;
 	}
 	
 	@Transactional
 	@Override
-	@CachePut(cacheNames ="employeeCache")
 	public DepartmentDTO addDepartment(DepartmentDTO departmentDTO) {
 		 Department department = new Department();
+		 logger.info("Running Thread from Department Service: "+Thread.currentThread().getName());
 		 mapDtoToEntity(departmentDTO, department);
 		 Department savedDepartment = departmentRepository.save(department);
 		 logger.info("Adding Department details into the database");
@@ -69,26 +72,51 @@ public class DepartmentServiceImpl implements DepartmentService
 	
 	@Transactional
 	@Override
-	@CachePut(cacheNames ="employeeCache",key="#deptId")
 	public DepartmentDTO updateDepartment(Integer deptId, DepartmentDTO departmentDTO) 
 	{
-		 Department department = departmentRepository.findByDepartmentId(deptId);		 
-		 Department dept = departmentRepository.save(mapDtoToEntity(departmentDTO, department));
+		 Department department = departmentRepository.findByDepartmentId(deptId);
+		 if(department==null) {
+				throw new IDNotFoundException("Department id not found  "+ deptId);
+			}
+		 Department dept = departmentRepository.save(mapDtoToEntityUpdate(departmentDTO, department));
 		 logger.info("Updating Department with Department Id: "+deptId);
 		 return mapEntityToDto(dept);
 	}
 
 	@Transactional
 	@Override
-	@CacheEvict(cacheNames ="employeeCache",key="#deptId")
 	public void deleteDepartment(Integer deptId) 
 	{
-		Department department=departmentRepository.findByDepartmentId(deptId);			
+		Department department=departmentRepository.findByDepartmentId(deptId);
+		if(department==null) {
+			throw new IDNotFoundException("Department id not found  "+ deptId);
+		}
 		EmployeeDepartment emdt=employeeDepartmentRepository.getById(deptId);
 		employeeDepartmentRepository.delete(emdt);
 		logger.info("Deleting Department of Department Id: "+deptId);
 		departmentRepository.delete(department); 
 	}
+	
+	@Transactional
+	@Async(value="taskExecutor")
+	@Override
+	public DepartmentDTO findDepartment(Integer departmentId) throws InterruptedException
+	{	
+		  
+		  Department department=departmentRepository.findByDepartmentId(departmentId);
+		 
+	      if(department==null) 
+	      {
+				throw new IDNotFoundException("Department id not found  "+ departmentId);
+		  }
+	      Thread.sleep(5000);
+	      DepartmentDTO deptdto = mapEntityToDto(department);
+	      logger.info("Fetching data from database");
+	      logger.info("Thread Name of Department Service:  "+Thread.currentThread().getName());
+
+		  return deptdto;
+	}
+
 	
 	private DepartmentDTO mapEntityToDto(Department savedDepartment) 
 	  { 
@@ -125,9 +153,28 @@ public class DepartmentServiceImpl implements DepartmentService
 			return department;
 			
 	  }
+	 private Department mapDtoToEntityUpdate(DepartmentDTO departmentDTO, Department department) 
+	  {
+		 department.setDepartmentId(departmentDTO.getDepartmentId());
+			//employee.setEmployeeId(employeeDTO.getEmployeeId());
+		 department.setDepartmentName(departmentDTO.getDepartmentName());
+		 department.setDepartmentLocation(departmentDTO.getDepartmentLocation());
+			
+			
+			EmployeeDepartment empDept = new EmployeeDepartment();
 
-
-	
-
+			
+			  List<Employee> emp=departmentDTO.getEmployee();
+			  
+			  for(Employee e:emp) {
+			  
+			  empDept.setDepartment(department); 
+			  empDept.setEmployee(e);
+			  //department.addEmployeeDepartment(empDept); 
+			  }
+			 
+			return department;
+			
+	  }
 
 }
